fix error in checking class of features or df argument
Checking the class of a matrix storing feature or pySENIC motifs with the class() function returns a length 2 vector "matrix" "array" that cannot be compared to 'character' using == without throwing up an error that prevents create_domino from running. The inclusion of [1] to only check the first element of the output of class() allows the lines to accomplish their intended function of checking if the arguments are file paths to read in results.