create_rl_map_cellphonedb initial commit
Creates a separate function from create_domino for the reading of a cellphonedb database so that signaling database inputs to create_domino can be of a generic rl_map dataframe format that can come from other rl database sources. Initially created without changing create_domino or fixing bugs with using complexes so that create_domino can be preserved until this function is complete.