return rl_map
After consulting some other lab members, storing the full tables from cellphoneDB is unnecessary and convoluted. Thus, the function just returns the data frame of the receptor-ligand map. Lines 27-33 address a bug with cellPhoneDB v2.0 where some proteins do not have some of their location or physical characteristics labeled and just have a "" in the corresponding cells instead of "True"/"False". These are filled with "False" to make sure the syntax check in the following lines work properly. The rl_map columns are also renamed to more clearly state receptor and ligands. There is also a column added to annotate the database used which could be used to write in version, but the user has to know what version they are using. An auto check would be best if I can find if its stored in any of the tables.