complexes maintained in rl_map
Following the formatting of CellPhoneDB cell-cell communication results supplied to the InterCellar app on github, the rl_map has been converted to a data frame with columns for feature names, gene names, uniprot_ids, and type as a ligand (L) or receptor(R). The function can read CellPhoneDB reference data for v2 or v4 with inclusion of interactions with complexes or not and support for gene conversion to orthologs. Complexes are no longer split into individual rows for each gene in the complex. The cell containing the genes comprising the complex is formated as a comma-separated list that will be parsed in create_domino to calculate correlations for each component and aggregate them for TF linkages.