gene ortholog fixes
ORTHOLOG CONVERSION ADDRESSES CELLPHONEDB'S LACK OF NON-HUMAN INTERACTIONS AND IS NOT PREFERRED OVER USING A REFERENCE SPECIFIC TO THE SAME ORGANISM THE SCRNA-SEQ DATA IS FROM. Uses the conversion dictionary from biomaRt to convert from the gene names used in the reference database to orthologs in the organism source of the scRNAseq data. If no ortholog for an individual gene or a gene in a complex is found, the gene is flagged, reported by printing to console, and the interaction is not included in the resulting rl_map. If a gene has multiple orthologs, those orthologs are included in the gene_A or gene_B column with semicolons (;) separating orthologs for the same gene.