create_domino reads rl_map data frame
use_complexes = FALSE does not work; intermediate commit to switch to approach of discarding complexes if use_complexes = FALSE. Unique gene conversions in create_rl_map_cellphonedb implemented to avoid redundant gene names in semicolon-seperated lists of orthologs. create_domino now takes the rl_map data frame from create_rl_map instead of filepath to downloaded cellphoneDB data files. Parses the rl_map into an rl_reading format where there is a consistent ordering of the receptor and ligand columns for interactions instead of partners A and B where either could be the receptor or ligand. End result is the list format of linkages where list names are a receptor gene or receptor complex name and the vector contains the  ligand genes or the ligand complexes that can interact with the receptor. The map of receptor-ligand interactions is stored in dom@misc. The hard-coded rec_min_thresh threshold is corrected to use the argument provided to the function (default expression in 2.5% of genes). First implementation of correlation between receptor complexes and TF activity using the median of the correlation coefficients between all components of the complex.