use_complexes FALSE ignores interactions
Shift in how the use_complexes argument is treated. Setting to FALSE now leads to all ligand-receptor interactions involving a protein complex as either partner are ignored. All genes encoding receptors or components of receptor complexes are assessed for expression correlation with TF activities. This initial, single gene correlation matrix is stored in dom@misc$rec_cor. The median correlation of genes in receptor complexes and the correlations of single-gene receptors are then compiled in dom@cor. This was done for user interpretability as the alternative approach was to separate genes in complexes into their individual components and establish linkages with those individual genes. This is how domino currently runs as of v0.1.1, and leads to issues where non-receptor genes belonging to a receptor complex undergo the correlation test and linkage to transcription factors regardless of whether or not the other receptor components are linked. This is also problematic for biological interpretations of the data as many of these complexes are non-functional unless all their constituent components are expressed. Thus, if a user sets use complexes to FALSE, their results will not include signaling interactions involving complexes. Users may be confused that important/expected interactions do not appear with this approach. Our current recommendation in such a context is to set use_complexes=TRUE or append your rl_map with the interactions the user is missing in a "'single receptor gene' & 'single ligand gene'" format.