Coord width scaling for high expression
For ligands with mean expression greater than the circos arc width of 1, a new scaled.mean.expression column in signaling_df is used to scale the expression values between 0 and 1 to allow comparisons of relative ligand expression across cell types. The use of this scaled measure only takes place if there is an expression value greater than 1 being plotted to keep low expression clear to the user if, for instance, the maximum expression value being plotted was 0.2 and we did not want that to appear as the maximum chord width of 1.