variable name for expressed ligand check
lig_genes stores the ligands present in the z-score matrix. inc_ligs has all ligands capable of interacting with the receptor from the rl_map. The use of the previous variable name led to errors running build domino, possibly because some ligand genes are not present in the z-score matrix.