legend includes chord width
Incorporates a panel to the legend showing how the width of each chord corresponds to mean scaled expression of the ligand from each cell type in terms of the minimum and maximum possible chord widths. The default scale is 0.01-1, but the upper bound will increase depending on the maximum expression value being plotted.