create_regulon_list_scenic
Define a separate function for generating the tf_targets list from a pyscenic grn function output. create_domino now takes a list formatted with transcription factors as names and the stored values being target genes of that TF's regulon. The create_domino function now takes this list as an optional parameter "tf_targets" to save the list in the linkages of the domino object. Params for gene conversion are removed from the definition of create_domino