Domino2
Reference commit for development update to Domino2 (v0.2.1) where updates addressing outstanding bugs in the initial publication of Domino, considerations for ligands and receptors that act in complex, uniform formats for data supplied to Domino, and new plotting functions have been implemented. Analyses conducted with this version of the package can refer to this commit.