presentation of loomR as alternative
Because loomR is not hosted on CRAN or BioConductor, trying to install the package during R CMD BUILD fails when trying to recreate our vignettes. Therefore, loomR use has been moved to a code block that is not evaluated and presented as a preferred alternative to using csv or tsv formats to save the counts matrix.