R scripts fixed to master branch version
There are many changes to the scripts in the R directory. Most are just indentation changes, but I cannot tell if there are also functional changes intermixed when almost every line is marked for changes. To be careful about not introducing new unintended changes beyond the pkgdown documentation, the scripts were fixed to their state in the 0.2.2 master branch