select_cluster_tf unit test
Simulates a matrix of 5 features for 300 cells belonging to 3 clusters. Feature scores are generated following a poisson distribution with the possibility of modeling the effect of belonging to one of the 3 clusters on the feature score. Scores are scaled to be between the values of 0 to 1. Tests are based on the modeled effects of a score being higher in one cluster compared to the others and that scores modeled with no cluster preference do not have a significant p-value difference between clusters.