assess_complex_receptor_cor initial
Using the list of receptor names parsed from the rl_map, reads through the correlation matrix to aggregate the measures of correlation used for receptor-TF linkage to identify component receptor genes for complexes and aggregate their measure of correlation. The default method first implemented is the median of correlations across all component genes of the complex.