select first gene in multiple uniprot mapping
Switch from sapply to vapply uncovered a bug with writing the rl_map where cases in which there are multiple genes associated with a uniprot ID in the CellPhoneDB genes table result in a simplified vector of all the genes being returned in the loop. By setting expectation of a length 1 character vector, we must enforce that a single gene name is returned. We will default to using the first orcurance of a mapped gene. If a more specific solution is required, its up to the user or CellPhoneDB to remove outdated or incorrect uniprot to gene mappings.