listFiles for bioc branch action
The prior approach to listing the files did not return an array to FILES (line 6). The new approach immediately lists all files in relevant directories for the bioconductor branch. Files are checked out one at a time in a for loop. Adds bin/bash to the start of the script.